2023 05, crystal, ~ per
The ruby version went on for hours. This is the crystal port. It still
took a few hours, but finally finished.

This version is the basic brute force algorithm. I'm not proud of it. In
a different timeline of my life I'd implement this with a range mapping
instead of single values.